Was able to get help to make testing work. ChessBoard: Added functions to set the chess board with pieces and added equals and hashCode overrides, I can't remember if I just fixed them or made them
ChessMove: Either added or fixed equals and hashcode override

ChessPosition: Added equals and hashcode override

Queen_Moves: made no_go a list of positions to not go past for values towards the bottom where i will be able to see if a piece is in the way before I add it to the list of valid_moves. Made it so the queen can move to one position before a piece of the same color or on top of a position of a piece of another color. Got rid of unessesary removals from bottom since I can prevent those from ever entering the list.