ChessGame: updated checkValidity() parameters, added booleans to check if black and white are either in check, in a stalemate, or in checkmate, maybe finished isInCheck(), inInCheckmate(), and isInStalemate(). Still not quite sure how to check if a team is in a stalemate. Currently planning on updating the positions of pieces in makeMove. Keep making new private variables so will wait to make overrides for equals and hashCode till the end.
ChessRules: I think i've got the parameters right at the moment. The params are the moves the piece can do, the positions of the teams pieces, the opposing team's pieces, and they type of piece being moved. For checkValidity() to work i'll make a Map in ChessBoard that gets filled as the board is getting set up. Rough idea for checkValidity() is for each possible valid move check if every enemy piece can put the king in check or checkmate with that piece being moved and keep track of valid and invalid moves based on that.