ChessGame: made rules final so it can't be changed
ChessRules: Working on checkValidity(). Realized i may need board to temporarily remove piece whose moves are being checked to see if opposing pieces can reach king if it's gone. Think I've generally got the gist of how to check if moving the king, if it's the piece in question, will put it in check. Working on checking if moving other pieces would put king in check.