ChessGame: decided I only needed the same color king's position for the piece whose moves im checking and not all other team pieces to check if enemy pieces can attack king if piece makes a certain move.
ChessRules: For checkValidity() updates its params to take king position and not the entire teams pieces. Originally was trying to make it so that i could check the validity of moves whether the piece moves belonged to a King or not. Decided, at least for now, that doing this probably wouldn't be worth it as for pieces other than the king I need to temporarily alter the board.