Pawn_Moves: Switched to my programming exam version cause it felt cleaner and my old one was causing issues.
ChessBoard: Just adjusted the newline character in my broken toString override.

ChessGame: Realized i never put anything in to get rid of the piece in the start position of the move after the move is made. Also changed the find king position loops to the top so no null king position values would get entered. Also made a check for if there was no piece at the start position.

ChessRules: Made it so for loops don't have to go all the way through before breaking if a move is bad. Did some more checking for using == and not .equals in createsCheck().