ChessPiece: Got rid of id because it was causing equals issues.
ChessBoard: Switched Map to Key = ChessPosition and Value = ChessPiece so pawns dont override each other and equality will still work.

ChessGame: Switched every Map from ChessPiece, ChessPosition to ChessPosition, ChessPiece. Finished goodMove() by checking to see if the move was one of the valid moves provided by checkValidity() so now all MakeMove tests pass. Made a helper function to get the kings position since I kept repeating that code. Started adding code to update if any team is in check, checkmate, or a stalemate when a new  board is given. Made two helper functions, one to see if a move puts a team in check and another to see if a team is in checkmate or stalemate. Working on implementing these in ChessRules now.

ChessRules: Realized that if I don't actually change the board my PawnMoves won't see that it can put the king in check so started implementing that.