ChessBoard: Fixed the formatting in the toString() override a bit.
ChessGame: Made it so that a new ChessGame automatically resets the chessboard. Added a break so that goodMove() wouldn't have to go through all of the valid moves after finding the move was in there. Changed goodMoves() so that after each valid move is made both colors are checked to see if they're in Checkmate or Stalemate. Made it so that setBoard() checks to see if the current color is already in check before it can override it by checking a new piece and fixed incorrect input values for endingChange(). Fixed color logic in endingChange(). Passed all tests!!!!!