Service Classes: Decided to make three service classes, one for user requests(register, login, logout), one for game requests(create game, join game, game list), and one to clear the database.
Handler Classes: Decided to put handlers in their own classes using same division as service classes. May change this and put all in one class depending on average size of handler. Decided to make a class to turn json to java and java to json for requests and responses.

Server: Currently dismantling handlers in there but will keep before authentication.