DataAccess: Added methods for dealing with games and implemented getGameList()
GameAccess: Ditto

GameHandlers: Added general imports and constructor variables. implemented gameList()

GameService: Made GameService and implemented getGameList()

InAndOut: Made getGameList() its own Java to Json converter for its response and made it so that ctx gets returned to handlers.

Server: Made it so context gets returned after its been modified. Added InAndOut as a class variable and added endpoint for gameList()

UserHandlers: Made it so InAndOut is passed in as a class variable. Made it so that response ctx is passed back to server.