Accesses: Added methods to each access to be able to get their current mock databases
DataAccess: Changed it so clear has a method to clear each database

GameHandlers: wasn't able to see that the context had changed so instead i made it check if the status code was anything other than 200 to see if there was an error.

GameService: made it so that addGame checks to see if the game name provided is null and made it so that whiteUsername and blackUsername have their initial value set to null. Also, added a check to see if color provided wasn't white or black and changed check from seeing if white or black username was an empty string to seeing if they were storing null

InAndOut: Realized I hadn't changed the error condition into json using gson and fixed that.

Server: Consolidated some imports and made it so that all of my accesses use the same instance of DataAccess

ServiceTests: Made service tests and checked they all worked

UserHandlers: Changed capitalization and added "Error" into my error messages

UserService: Realized that error message in registerUsers() try block was necessary and wasn't needed because one of my data access methods was broken. Also added username as return value for successful result in registerUser() and loginUser(). removed unnecessary variable to store return values i didn't need