Server: Hopefully, made it so that mysql is used and saved as soon as the server initially starts.
Clear: Made it so that errors and exceptions are handled

ClearService: Made it so that it can throw exceptions and errors

GameDatabaseAccess: Started to add functionality to clearing game data and adding games.

UserAccess: Made it so that clearing the db and getting the users from the db can throw errors and exceptions.

UserDatabaseAccess: Hopefully added functionality to clearing the user database, getting a list of all user info, adding a user, and getting the info for one user.

Added files for everything else